Fix remote editor deserialization once and for all
Summary:
We currently rely on Atom's broken but functional flow for deserializing remote editors:

1) Atom serializes the TextEditor with the `nuclide://hostname/x` URI.
2) During deserialization, Atom's TextBuffer attempts to reload `nuclide://hostname/x` on local disk. At some point, it normalizes the path, so it becomes `nuclide:/hostname/x`.
3) We detect all broken `nuclide:/hostname/x` URLs after reconnection and reload them with a correct editor, which has a `RemoteFile` backing it.

In Atom 1.22, this no longer works because step 2) got stricter, and non-existent files are automatically closed. Bad news for us! Also, we have to account for broken `nuclide:/` URIs all throughout Nuclide which is super frustrating and leads to bugs. On Windows, this breaks Atom from loading at all! (I had patched the deserializer to fix this; that patch can now be removed).

To fix this problem once and for all, matthewwithanm had the idea of overriding the TextEditor serializer to deserialize into a placeholder `RemoteTextEditorPlaceholder` pane item (under our control). This allows us to control the complete remote reload experience, without relying on Atom quirks.

The serialized `RemoteTextEditorPlaceholder` contains the URI, contents, and modified state. After deserialization is complete, this creates a read-only text editor with the stored contents and also a user-friendly notice indicating that the file will be restored after reconnection. After reconnection, we still have to destroy and replace the placeholders with real editors, but this becomes much more reliable now that we're not relying on looking for broken paths.

In the future, we can even do things like add buttons to bring up the reconnection dialog.. but this is already a huge leap in user experience.

After a few weeks of having this out, I plan to modify the code to immediately kill all broken editors and clean up all of Nuclide's handling around broken URIs. (T23250374)

Reviewed By: mostafaeweda, matthewwithanm

Differential Revision: D6219051

fbshipit-source-id: 26ed1c7a47806a8ca1b6c8fdff3308f812affd6c