Fix file-tree rendering nodes that should be invisible
Summary:
The method `FileTreeNode.findByIndex` was prone to returning a node that is in fact not visible in some (not very common) scenarios.
This would happen when we'd continue searching from a `this.nextSibling` not testing whether this `.nextSibling` is in fact visible or not.
Returning such node would mess the calculation of file-tree's indexes of shown rows and would cause janky scrolling

Reviewed By: matthewwithanm

Differential Revision: D6237034

fbshipit-source-id: 9bfa7dd92423dfafb1a2124741e6969aaacb52e0