Make vscode-react-native independent of vscode
Summary:
vscode-react-native extension starts a server for each React Native project opened that manages the packager lifecycle and provides hooks to show alerts and open files in the VS code UI. Launching a debugger spawns a client that asks the server to launch through the packager.

This commit skips the server spawning step and has the debug entry point launching the packager directly and handling debugger communication.

Reviewed By: mostafaeweda

Differential Revision: D6188627

fbshipit-source-id: 6674291f8b870e465de48b13aac58ff12d76a906