command to see in-scope variables
Summary:
Enumerate scopes from the most recent stack frame, and display a combined set of all variables in those scopes, leaving out scopes
that are marked as expensive by the adapter (likely to be very large). In an upcoming diff the user will be able to manually specify to see
the content of those scopes, and to pick stack frames other than the default.

Depends on D6225524

Reviewed By: mostafaeweda

Differential Revision: D6230033

fbshipit-source-id: 27c9681fb2cd70a23272f610e263a15bb1eb48a7