Add input validation state to AtomInput
Summary:
Adding a validation state to atom input so that users of the component could inform the user of input errors. The styling is kept consistent with the atom input/form html element invalid states (see style guide). Also deleting custom styling on the RegExpFilter component.

Note that find and replace has a different style. This could potentially be a inconsistency with the theme.

Reviewed By: matthewwithanm

Differential Revision: D6264466

fbshipit-source-id: cc37de652af1ecf2bd15fa46f421850453c4fa4e