Wait until a full poll finished to call the next one
Summary:
Instead of fetching every T seconds, which is error prone because sometimes adb is dumb slow, let's just wait until every fetch finishes in order to call the next one. There's at least a 1 second interval between fetches.
It took me ages to figure out why polling randomly failed, and it was because of this.

Reviewed By: ebluestein

Differential Revision: D6266586

fbshipit-source-id: cc01ca716aa6b308f6e5eb6690dd0e6fa2749fef