Reduce the amount of ps memory checks
Summary:
Currently, to get the memory usage of Clang processes we run 'ps' for each process.
Unfortunately `spawn()` can block the main process, so having this many spawn calls is a huge perf hit :(

Instead of calling `ps` for each process, we'll instead aggregate the PIDs and run one `ps` with all of them!

Reviewed By: a20012251

Differential Revision: D6281712

fbshipit-source-id: 76f9b361ac77e2cd67074efc7792ef0ef1bd13e3