Fix LRU cache-vs-promise problem with FlowExecInfoContainer
Summary:
I was mystified about why we were allocating 40KB of FlowExecOptions objects:

https://pxl.cl/9FdP

We don't set the item in the LRU cache until we finish awaiting the promise. This is a big problem: if you are flooded with calls to `getFlowExecInfo`, they'll all simultaneously start a new promise and all await their own promise. Only one will eventually get set in the cache.

Instead, we should immediately store the promise in the LRU cache and return it. This is already an async function, so the caller can `await`. And it's fine to await a promise multiple times, by my understanding.

Am I missing anything?

Reviewed By: nmote

Differential Revision: D6274345

fbshipit-source-id: 67390d500f218208810830c56eb2bc64aad52a5c