Any type delegate mock
Summary:
This hasn't been blowing up in Sandcastle, or locally but I think it should. Further, if you touch nearby code and then return it to the current state (master), Flow will get into a state where it _will_ start to realize this is an error.

I'm guessing this is a Flow bug or something. `manager._linterDelegate` is typed as an instance of `?IndieLinterDelegate`. Therefore I would expect this line to always fail a Flow check since `manager._lineterDelegate` is typed as `?IndieLinterDelegate` and the `delegate` POJO is not actually an instance of `IndieLinterDelegate`.

Reviewed By: matthewwithanm

Differential Revision: D6287500

fbshipit-source-id: 0eba9e7bc78e9085d70abf16e2b14e1c08b411b0