Create wrapper for Relay QueryRenderers
Summary:
This component will display a loading icon while the QueryRenderer is querying results.
If there is an error, it will send out an atom error notification with the error message.
If there is no error, and the results are loaded, it will call the render function you pass it, with the QueryRenderer props as arguments.

Reviewed By: wbinnssmith

Differential Revision: D6279657

fbshipit-source-id: 67d604a9e58e323b28ff4e35b2d3d680b413a3a8