JS imports: Overhaul the 'package main file' resolution approach
Summary:
We used to identify files that were "main" files (as defined in a `package.json`) by crawling up from each file for a `package.json` file and reading the contents. This is unnecessarily slow - most projects have a small amount of `package.json` files so it's better to list them all up front and retrieve a mapping of main files.

This diff does exactly that. We still need the old "main-finding" for file changes so I've kept it around. I noticed that it was slightly inefficient before so I rewrote it with `lodash.memoize`.

Reviewed By: wanderley

Differential Revision: D6291719

fbshipit-source-id: 266f1a79950cbd096ae20f86587633b89cbc9efa