JS imports: Cache the list of exports for every file on disk
Summary:
This uses `DiskCache` to store the list of exports for all JS files, loads it at startup, and persists it after indexing is complete. The export cache is included in the `FileIndex` object introduced in a prior diff.

For `node_modules` files the caching is a bit tricky, since we don't know the entry points up front. I could cache this as well, but since the number of node modules isn't very large it's fast enough to still look up the entry points at run time. (and the exports are still cached for each entry point).

Some subtle points:

1) The cache is keyed with the full path + serialized haste settings, in case that changes. There's also a "cache version" which should be bumped if there are bugfixes or similar (ideally this would be auto-generated based on the server code, but...) I didn't want to use the Nuclide version as this usually won't have to change.
2) A new cache is created for every run so we won't keep cache entries for deleted files around forever.
3) Caches are keyed off the SHA1 hashes provided by Watchman when we first list the files! We compute the hash of new files inside `getExportsForFile`.

Reviewed By: wbinnssmith

Differential Revision: D6291716

fbshipit-source-id: f25500f2a26350ec593cea9a01777f25b9f22dab