Add SafeStreamMessageReader to catch jsonrpc exceptions
Summary:
`vscode-jsonrpc`'s StreamMessageReader has a fatal flaw in that it throws uncatchable exceptions on stream errors - which is weird because it does have a facility for emitting errors back to the user.

To work around this I've created a custom `SafeStreamMessageReader` that overrides `onData` to catch any exceptions thrown by the parent `onData`, e.g.:

https://github.com/Microsoft/vscode-languageserver-node/blob/master/jsonrpc/src/messageReader.ts#L220

Upon receiving an exception we'll emit an error via the usual means and tear down the emitter to prevent processing of further events. I've also modified all the existing users to use `SafeStreamMessageReader`.

This does rely heavily on implementation details but the unit test should have us covered if we upgrade the dependency.

Reviewed By: arxanas

Differential Revision: D6300352

fbshipit-source-id: 1f119d1c93ae96962c07d8066e85bf33fd8de356