Use promise to avoid races on _managedRoots field
Summary: Previously there were several awaits between checking the existence of the clang request setting and setting it in the map, so two calls to addClangRequest (e.g. two files under same buck target) could have set the same key and we would be left with a zombie clangd process that we could not reap on dispose.

Reviewed By: a20012251

Differential Revision: D6313777

fbshipit-source-id: 95e5e268d12ca063a552a6383266c49f04179317