Lazily load client-side proxies
Summary:
On the client-side we always explicitly `getServiceByNuclideUri` so it's easy to implement lazy-loading of the service proxies.
There's also no way for the server to call client-side methods without explicitly having something from the client.

This helps chip away at one major blocker to enabling the RPC framework for local development, as the loading cost is quite high.
This should also decrease load time for remote projects in general, as less code gets loaded on startup even in production :)

Differential Revision: D6268525

fbshipit-source-id: fa9d4fc5b4a5b0da173d2d32a833ce938adf9ee5