Cache parsed ASTs and generated proxies during development
Summary:
The biggest blocker preventing us from always enabling RPC for local services is the startup time penalty in dev mode.
I'd like to enable this, even in production, to make it easy to offload RPC modules (e.g. Flow, Hg) into a separate process.
Lazy loading (the previous diff) basically eliminates the cost in production, but in dev mode the reload time hit is painful :(

After profiling, the most expensive costs are actually just loading the additional dependencies (babylon, babel-types, babel-traverse)!

This uses temporary files to cache the results of the two most expensive parts: parsing the RPC definition files and generating the proxies afterward.

This leads to a rather unusual caching pattern as both the function and its dependencies are expensive to fetch, and yet we need to invalidate the cache if any of those dependencies change. `rpcCache` takes a function along with its arguments and an "argument key" - please let me know if there are any better/clearer ways to write this :)

Reviewed By: rightparen

Differential Revision: D6268526

fbshipit-source-id: 97e98f305b216767f54a5fef6bec4d9d132f44f8