Add a flag to start local RPC services in a separate process
Summary:
With the preceding diffs, it's possible to now:

1) Enable `useLocalRpc` without any performance penalty, even in dev mode
2) Start adding a flag to start RPC code in a separate local process.

We've had many reports lately of users experience UI stalls when working in local repos due to `spawn()` unexpectedly blocking.

It's actually not hard to address this systematically by just moving all RPC activity into a separate process - so Flow, Hg, etc would all be spawned by this separate process instead of blocking the main process. Nuclide then communicates with this separate RPC server using special Node-IPC based transports.

I've added an experimental `useLocalRpcServer` flag so that we can test this behavior for a while before enabling it more widely. My rollout plan is roughly:

1) Enable `useLocalRpc` for all Nuclide devs, but have server mode be opt-in
2) Enable `useLocalRpcServer` for all Nuclide devs
3) Start enabling `useLocalRpcServer` for beta testers

There are still various code fragments in -rpc modules that hackishly use the `atom` global if it's available, so those are going to have to be addressed first.

Reviewed By: rightparen

Differential Revision: D6268523

fbshipit-source-id: 89734bbac2f3966485556e0370db3cd382f7aef1