Fix TestRunnerPanel.js
Summary:
All changes were performed manually if trivial and if there would be no behavioral change. Otherwise, the new lint rule was suppressed.

It likely had a state value removed, or it used state in a risky way such as passing the entire state object around or spreading it. If the latter, it may have had an `eslint-disable` pragma added to suppress warnings, or -- if trivial -- modified to make passed state explicit.

This component may also have:
* Had constructor assignments moved to property initializers if the constructor had trivial assignments to initial state and/or props.
* Had explicit types for props/state class properties removed. These are no longer needed as Props and State types are provided as arguments to the polymorphic type `React.Component`

You're a reviewer on this diff because you were recommended as one :)

Reviewed By: matthewwithanm

Differential Revision: D6292701

fbshipit-source-id: 045a6f403f71a13d599ca9b88893e08251bd0d94