Fix ConsoleContainer.js
Summary:
This file was involved in the introduction of react/no-unused-state.

All changes were performed manually if trivial and if there would be no behavioral change. Otherwise, the new lint rule was suppressed.

It likely had a state value removed, or it used state in a risky way such as passing the entire state object around or spreading it. If the latter, it may have had an `eslint-disable` pragma added to suppress warnings, or -- if trivial -- modified to make passed state explicit.

This component may also have:
* Had constructor assignments moved to property initializers if the constructor had trivial assignments to initial state and/or props.
* Had explicit types for props/state class properties removed. These are no longer needed as Props and State types are provided as arguments to the polymorphic type `React.Component`

You're a reviewer on this diff because you were recommended as one :)

Reviewed By: matthewwithanm

Differential Revision: D6292705

fbshipit-source-id: 7e78ecb22a51474335dee87e257eb48093f0db6f