Replace isRunningInClient with isRunningInServer
Summary:
After the change to run local RPC processes, the definition of "client" just got a lot more fuzzy - is it in Atom, or is it in the local process? It's no longer sufficient just to check for the Atom case.

I've renamed `isRunningInClient` to `isRunningInServer` to make this distinction clearer - for analytics purposes we should still preserve the meaning of "client" as something running alongside the Atom process.

Many of the existing callsites really mean to say "if we're running in Atom", so this makes that situation much clearer.

Reviewed By: rightparen

Differential Revision: D6320115

fbshipit-source-id: 2c9ad1139d8d688df2b69d2356d0e8f47a4f3eb3