Improved UI for changed files in file tree
Summary:
The hover over action buttons of the changed files list (eg. diff view, revert) will always be visible on hover now, and if the window is too small, the filenames will be truncated with ellipsis. Also the action buttons will always appear anchored to the right side of the pane.
This change will affect the changed files list in the file tree, diff view, and commit form. However, the interactive smartlog will not have the action buttons anchored to the right, as the window is usually very large, so it doesn't make sense.

Reviewed By: shushz

Differential Revision: D6222040

fbshipit-source-id: b1e1b6a561464ee8d7e25eca7416b595a74bed95