bind {alt,ctrl}-backspace in terminal
Summary: On (Windows/Linux | Mac) it is common to use (ctrl | alt/option)-backspace to erase the word preceding the cursor.  In readline (used by bash to read commands), the sequence ESC BACKSPACE erases the word preceding the cursor.  evangrayk made this work in D6056473, this just adds a default keybinding to enable it out of the box.

Reviewed By: evangrayk

Differential Revision: D6339446

fbshipit-source-id: aa25635e53b43df836e49b858458f2030f26e8cb