Add protocol logging for QueuedAckTransport
Summary:
We have found that QueuedAckTransport introduces an issue where an object ID is unrecognized on some machines.  This adds logging to keep a circular log of every message sent via QueuedAckTransport along with every object id that is registered/unregistered locally or remotely.  Message bodies have common protocol strings removed, are truncated to 140 characters, and are JSON-string encoded.  We also remove the payload of any 'writeInput' method so that we don't accidentally record a typed password.

In a sample during startup, I saw 165 messages per second.  Given retention of 5 minutes, at 200 messages per second with 200 bytes per message, this is 12 Megs.  This is a substantial overhead, but even a factor of 2 or 3 above this seems tolerable for the small population running helping to test with this enabled.

Reviewed By: mostafaeweda

Differential Revision: D6282223

fbshipit-source-id: f83e1dd13574ee498760f0b24ca5636962244877