Update QueuedAckTransport protocol
Summary: This updates the protocol for QueuedAckTransport to assume that the underlying transport is generally well-behaved except that it can disconnect when the network is interrupted.  By making this assumption, it can get away with resending messages only on reconnect.  It handles receiving messages out of order (which may happen during reconnect), and it sends acks every 100 ms when there is outstanding work.

Reviewed By: hansonw

Differential Revision: D6282985

fbshipit-source-id: b54d652de1d99ae956012cc8020730d924f94868