Handle file actions while server is shutting down
Summary:
If there's been an error, we enter the 'Stopping' state, do a shutdown request (awaiting up to 30s for a response), then send an 'exit' notification, then dispose of the lspConnection, then switch to 'Stopped' state.

It's possible that Nuclide actions might arrive while we're in 'Stopping' state! The code already handles that case fine for regular requests e.g. getDefinition which simply returns immediately if we're not in 'Running' state.

But the current code for fileOpen/Close/Save/Edit has an INVARIANT that we're 'Running'. This invariant is incorrect. We might be running or stopping.

This diff fixes up the invariant.

Reviewed By: arxanas

Differential Revision: D6327979

fbshipit-source-id: f21c54dd4e6afb2bd1bd13292c1f60ec4da7a9fe