Create DropdownResults component
Summary:
The Need:
The Combobox component is a little too smart and is really hard to use in specific scenarios.

This new component will just be a more stupid component, that will only know about a list of options it needs to show, and hide itself when it is not needed.

Reviewed By: matthewwithanm

Differential Revision: D6350774

fbshipit-source-id: 46a3b4d38eb3ee5481bd12a6ace490973517ead6