Add Application Info data support to device panel
Summary:
This set of diffs will result in the ability of displaying and edit the sandbox that is configured in Android Facebook app.
I'm going to do this by showing this information in the current devices panel that we have today in Nuclide.

This is what it looks like today once you chose a device to see information about:
{F115389735}

My plan is to add the concept of "Application Info" type of table into this view. It will be a table where we can show/modify information about running apps on the device.

This diff adds the necessary data structures and data flow that will allow any other package to specify how to display data on the "Application Info" tables.
The device panel package achieves this by defining the concept of "data providers" for the different types of data it shows. Any package can implement and register them with the device panel package. This way the device panel only cares about displaying data and not the different details of collecting it, that is left to other packages.

This diff defines the `AppInfoProvider` API and all necessary structure that allows other packages to create and add to it.
An `AppInfoProvider` needs to specify an type (which platform), a process name (basically the app), a name (the information name) and an observe function that will observe the information value for changes.

The information obtained through these providers will be grouped by the process name and each group will be rendered in its own table.

Reviewed By: stepanhruda

Differential Revision: D6391809

fbshipit-source-id: ec81503902a1d25b669767e37f241d55b17ab087