use default shell
Summary:
On Linux and Mac there are standard-ish mechanisms for specifying the default shell.  Nuclide terminal already had a mechanism for specifying the shell to use (~/.nuclide-terminal.json).  This makes the terminal use the following shell in order of preference:

1) Command specified by ~/.nuclide-terminal.json
2) Default shell (from `getent passwd $USER` on Linux, from `dscl . -read $HOME UserShell` on Mac)
3) '/bin/bash' on Linux/Mac, or 'cmd.exe' on Windows

Previously we were passing `--login` and `-i` to bash.  However, we cannot generally know these kinds of options for arbitrary shells.  The `login` program handles this by prepending '-' in argv[0] while calling `exec` with the actual path to the shell, but 'node-pty' does not support this.  Since most terminal programs don't start login shells anyway, I'm omitting arguments by default.  It is still possible to use '~/.nuclide-terminal.json' to add them.

Reviewed By: mostafaeweda

Differential Revision: D6365906

fbshipit-source-id: d73aa248eb8534d9d266892fd3c9da120df77c50