Merge useLocalRpcServer into useLocalRpc setting
Summary:
This enables useLocalRpcServer whenever just useLocalRpc is activated, to simplify this code quite a bit.

This doesn't really have to be done in the startup path anymore - it can be done the first time a service is requested. (This isn't an optimization at all, though, since there are several packages that start using RPC services at load time -- although they probably shouldn't.)

Reviewed By: rightparen

Differential Revision: D6413250

fbshipit-source-id: c90dbd6fedffb97499ef776f629caa984333771b