Enable the local RPC process with the nuclide_local_rpc gatekeeper
Summary:
This controls the local RPC process with a nuclide_local_rpc GK.

Note that it's not possible to use an asynchronous GK query here (since RPC services are expected to be available immediately). It would be possible to defer the asynchronicity by queueing up the transport messages, but if the GK fails then we would pay the cost of marshaling without any benefits, which isn't really acceptable :(

This means that the GK check actually relies on the cached value from the /previous/ session. Unfortunately this means that the setting won't actually apply until the *second* time the user reloads.. but this should be sufficient to get some reasonable testing done. I've also added a track call so that it's possible to get some data from users who do eventually pass the GK.

Reviewed By: semmypurewal

Differential Revision: D6413251

fbshipit-source-id: 939eac331465b4daf6914a76964113acc60641be