Custom arg parsing per adapter
Summary:
Some adapters take object parameters, and for these it's easier for the user if there's
a more convenient way to enter them than to try to put JSON on the command line. This diff allows
per-adapter custom parsers for any adapter parameter, which the command line parser will use
instead of the type specified in the adapter's package.json.

Initially, this is for --sourceMapPathOverrides on node.

Reviewed By: mostafaeweda

Differential Revision: D6389873

fbshipit-source-id: 1d0fbd7e5e462901523c93788081ca017d908ac5