Integrate React DevTools with Sonar
Summary:
Previously, React DevTools worked out of the box with React Native applications. The way we could run React Devtools was by doing `js1 devtools`.

This diff embeds React DevTools within Sonar so that product developers no longer have to launch the standalone Electron shell for React DevTools; they can use Sonar for all their native development needs.

Since React DevTools is designed to also function as a standalone application, it opens up a websocket connection from the desktop to the phone. The major challenge of this diff was to force React DevTools to leverage Sonar for its communiation instead of its personal websocket connection. This was accomplished by passing in fake websockets to the React DevTools client and server. On the surface, these fakes look like real websockets, but under the hood, they subscribe to the Sonar connection and relay messages on the Sonar port. To make this work, I also had to modify the API of React DevTools. So, I'll launch a corresponding PR on the public respository that enables the changes in this diff to work.

**Note:** Most of the work here depends on the files introduced by andrewimm in D5559753. In D6267163, I migrated the JavaScript `SonarClient` to `fbsource`. I also had to update React DevTools to v3.0.0 to get these changes to work.

Reviewed By: emilsjolander

Differential Revision: D6267245

fbshipit-source-id: 823adcde0a53918116dbe497843913d7875164c4