Add UI support to device panel to edit App Info values
Summary:
I again did two things in one diff:

**Moved `AppInfoValueCell` into its own file**:
It was growing quite a bit so I moved it.
**Added UI to allow edition:**
The new UI is based on two new props `canUpdate` and `update` function.
When a value provider can be updated we render it with an edit button shown at the right side. When you click on it you enter into a mini state machine that allows you to edit the value:
* **`editing`**: an input is shown and you can change the value, if you hit enter or get out of focus the data is `synced` to the provider.
* **`synced`**: the data is being synced with the provider and a little icon shows to present that state. this means the promise for the update hasn't been resolved or rejected yet. When it's resolved it is `updated` and if it's rejected it's an `error`.
* **`updated`**: the value was successfully updated and we show an icon to indicate that for 1s. We then finish the editing state.
* **`error`**: similar to `updated` but we show an error icon.

Reviewed By: stepanhruda

Differential Revision: D6403377

fbshipit-source-id: 74152c5cde5bb67f0e0620dc8650c16c36fce0d2