Small cleanup to diagnostics tooltip
Summary:
Remove an out-of-date comment and improve the semantics of the DOM node used for the gutter indicator.

Currently, because the gutter is an `<a>` it has a strange looking hover state (the icon gets underlined) and the cursor switches to a "pointer" which inacuratly indicates clickability. The hover state is masked by the tooltip which appears over it when hovered, but as we explore other behaviors for the gutter icon, this artifact may become noticeable.

Reviewed By: matthewwithanm

Differential Revision: D6425920

fbshipit-source-id: 60ce132d3f2b7a92931d4b81b194d5d48fe7a114