Show and record DOM counters in Nuclide health
Summary:
The Chrome debugger API allows us to gather a count of all retained DOM nodes and event listeners (including detached nodes):
https://chromedevtools.github.io/devtools-protocol/tot/Memory/#method-getDOMCounters

I've also added the count of attached nodes in the DOM to get a sense of 'leakage', although this doesn't directly allow us to find the sources.
This only runs every minute (because expensive), and so I also prevented it from being visible in the jewel.

Reviewed By: matthewwithanm

Differential Revision: D6429472

fbshipit-source-id: 554b5d3e5761c731858d9b85e4b10ee934207957