Get refactors by range instead of by point
Summary:
`nuclide-refactorizer` uses a point to get a list of refactors. Just a point is not enough to determine if refactors like "extract variable" or "extract function" are valid ones because they need to validate the selected range.

This diff changes `refactoringsAtPoint` to `refactorings` which receives a `Range` instead of `Point`.

Reviewed By: hchau-fb

Differential Revision: D6392526

fbshipit-source-id: 0d2053079c10cf4b8dfb0c25d8668cd74be51b17