Eagerly increment buffer version in BufferSubscription
Summary:
Because of the async delay before the event is actually sent, it's important to immediately increment `this._changeCount` before the `await`. (This is so that clients like autocomplete know to send the correct version).

There's also another issue where multiple changes must be sent over in reverse order, as edits at the beginning of the document affect edits later in the file.

I've added a spec to codify this.

Reviewed By: drmathochist

Differential Revision: D6440311

fbshipit-source-id: 6cf789f0d72b35bdbc74fa0187181cdfcbf018c5