Spawn local RPC process with --inspect flag in dev mode
Summary:
This allows easy debugging from chrome://inspect. The port is generated randomly; look in the Atom console to see which port was used (or `ps ax | grep LocalRpc`).

This also enables source maps to enable easy debugging with another Nuclide window! :D

One hacky bit is that I have to strip away the `--inspect` flag from the child process to allow it to continue forking -- otherwise the forked process will attempt to use the same port and die instantly.

Reviewed By: mostafaeweda

Differential Revision: D6441168

fbshipit-source-id: 41f03b91c58f71eeb5bb4c64de546e092e7f88dc