Re-factored appinfo epic
Summary:
Completely rewrites setAppInfoEpic.
The original one had a few problems that were brought up by matthewwithanm when I asked him about Observables.

For instance, the observable data flow set didn't allow me to apply distinct() on the provider values being observed in order to reduce the number of produced data through the observer flow.

Another issue solved is that value production doesn't depend anymore from the SET_PROCESSES action since each provider value is being observed on a timer basis that is independent from anything else.

Because of the better data flow here I was able to change the Table and Row UI components to be a PureComponent, this way we greatly reduce the number of renders as they were being rendered every 3s because that's the frequency of the processes update.

Reviewed By: stepanhruda

Differential Revision: D6438557

fbshipit-source-id: 377048a6b6984b180c08e79f76f9836a3414130f