Terminal should start login shell by default
Summary:
Nuclide terminal traditionally started `bash --login` by default.  When I changed it to support the user's default shell, I didn't want to assume that the shell supports `--login`, and the pty library we use does not support adding `-` to the shell's argv[0] name.  So I removed the `--login` and made the shell non-login by default.  This broke some people's workflow.

It is a bit ambiguous, but we have a precedent at this point for starting login shells.  Since we cannot change argv[0], we pass `-l`.  Bourne shell does not support `-l`, but every other shell I know of does, and I don't think Bourne shell compatibility is important here.

Reviewed By: kevinastone

Differential Revision: D6435150

fbshipit-source-id: ec968f7ba5116cd2110613b61fe9b4aa80d3a0ea