Generate unique token for terminal URIs
Summary:
Atom supports a protocol for opening external links.  I don't think there is currently a way this can propagate to the terminal code, but I'm adding this as defense in depth in case that is ever supported.

Allowing a website to control the commands that run in a terminal represents a security hole.  This ensures that terminals with any pre-filled commands given to a terminal come from inside Nuclide.  A side effect of this change is that any persisted terminals will rehydrate in a default state.

Reviewed By: hansonw

Differential Revision: D6415664

fbshipit-source-id: 09c34e81f74f8259b032231a7b1a222889bbf5d8