JS imports: use 'hg files' and 'find' before falling back to glob
Summary:
We've had some reports of the JS imports server running out of memory!

I verified by inspecting the Node core dumps that the OOM was actually coming from `glob` (yes running `glob` with 1M+ files will cause an OOM).

We have a lot of other things to try before resorting to glob - in particular `hg files` and `find`.
This also logs an error if Watchman listing doesn't work so I can try to find out why we even have to do this.

Reviewed By: pelmers

Differential Revision: D6455166

fbshipit-source-id: 613f6a45e7d389c0d775d6fd0c26e89cdc7d048f