Track local RPC process memory + CPU usage
Summary: Offloaded tasks are not tracked by the main Nuclide process, so we'll have to do it manually. CPU usage is tricky because it's cumulative -- we have to maintain a delta. In theory the API is supposed to take a delta given the previous value, but this only works for one iteration.

Reviewed By: rightparen

Differential Revision: D6472067

fbshipit-source-id: 546fcf9be4d13035be6bbc51c9b5e3943e348994