Fix memory leak in PathWithFileIcon
Summary:
This leak is much simpler than the previous: it's just due to the fact that the disposable returned from `_consumeFileIconService` is actually retained for use when `file-icons` is disposed. Since that's essentially forever, this means that `PathWithFileIcon` can never be collected (and since React components hold onto a lot of state, this generally also includes all containing components!)

The fix is straightforward: we can also trigger the disposable upon unmount.
I've also cleaned up the async consumption hack since the fix was merged back in July. (thanks Matthew!)

Reviewed By: rightparen

Differential Revision: D6472077

fbshipit-source-id: a27f1c61ba4d3638ed92d234156e2674e466b4e1