fix terminal resize issue
Summary: Evidently the xterm-viewport DOM node can change if you drag a terminal between panes.  This code has been through a few iterations and I'm not sure why we needed to look up the xterm-viewport anyway when the root div has the same dimensions and appears to work correctly with observeElementDimensions.

Reviewed By: mostafaeweda

Differential Revision: D6509836

fbshipit-source-id: da495eb720f1154f1b2303d644662bdba8df1708