do not assert if one role suffers amnesia
Summary:
There is a potential race where the client is about to reconnect when the user closes the connection.  The client sends InfoService.closeConnection, the server deletes its QueuedAckTransport for that client, and then the client reconnects.  The first thing the client does is to ack the last message it has processed.  Since the server sees the reconnect as a new connection, its QueuedAckTransport has no state so we fail.

This fixes the issue by making QueuedAckTransport robust to this.  There is nothing productive we can do at this point since we don't know if anything got dropped, so we track and close, but we should not crash.

Reviewed By: hansonw

Differential Revision: D6505306

fbshipit-source-id: 2edd4496c5db0712c9045b440d56be83c79e8cc5