JS Imports: Make sure that missing dependencies are found even when index is wrong
Summary:
When you open Nuclide, the exports is index has not actually populated yet. So AutoImportsManager will report no available exports, and set that in its cache. To get around this, we check the cache, but if it says there are no returns, we check again just in case.

Not the best solution, but there is no way to tell currently what the status of the export index is.

Reviewed By: hansonw

Differential Revision: D6450700

fbshipit-source-id: e040a9a164985443e62a4f78e40618c98d2364e3