Nuclide VSTranslator: Fix issues with stopThreadId
Summary:
This fixes a problem with the ThreadStore and ThreadEvents where if we add a new thread while the program is already paused (VS code ThreadEvent, reason = 'started', followed by a StoppedEvent with threadID = <new thread> and allthreadspaused = false to indicate the new thread is paused like the rest of the target) the thread appears in the thread list (desired) but it also takes focus as the active thread in the debugger UX (bad). This is problematic because it will replace the call stack and scopes frames and is extremely annoying to the user. Additionally, if Nuclide doesn't have focus, a native toast notification is raised saying "Nuclide has paused at a breakpoint..." It's terrible.

This fix sends a stop thread ID and a paused event only if the stop event comes while the program is not stopped.

Reviewed By: mostafaeweda

Differential Revision: D6342268

fbshipit-source-id: 11490a822d5362edcef98c9738d1ba6676054984