Use a provided project root when attaching from the device panel
Summary:
Sometimes, for a given symbol, lldb can tell Nuclide the relative path of the file containing that symbol.
When using the buck task runner, the project root is known and the lldb integration can correct prepend the project root to the relative file path provided by buck.
However, when using the attach button from the processor table in the device panel, no project root is known and, unless it is provided, those files couldn't be properly displayed.

Reviewed By: sas

Differential Revision: D6516302

fbshipit-source-id: 203b9c970209ba9fc8dc912ed8e3833b3ca53bea